+++ cmd/fireeth/tools_compare_block_store_rpc.go
@@ -97,7 +97,7 @@ func createCompareBlocksStoreRPCE(logger *zap.Logger) firecore.CommandExecutor {
 				panic(err)
 			}
 
-			receipts, err := blockfetcher.FetchReceipts(ctx, rpcBlock, rpcClient, 20)
+			receipts, err := blockfetcher.FetchReceipts(ctx, rpcBlock, rpcClient, 20, false)
 			if err != nil {
 				panic(err)
 			}